Committing my solo overtime work. I made changed in the following locations:
validators --> Changed the description validator to allow for up to 150 characters rather than only 40 (because the field allows for 150).

unit tests --> Removed some comments Megan and I put in for debugging (which doesn't make the unit test output look very nice)

kvp -->Changed the year selector to be in reverse, so that the latest year is the first one listed (it doesn't make sense to have 1980 listed first, and then force the user to scroll to find 2017)

templates --> Put all the templates inside a container class to make pages fit nicer on the screen. Changed how the buttons look (got rid of the hard 15% width on them), added some subjective wording depending on whether something is being created or edited, and I standardized font styles and sizes across the app's templates.

-- Cam

Total time spent: 60 min (1.2hr of projects time @ 50 min/hour)