Finished solo overtime.
Short version: Contact logs are now displayed in the case's edit page through HTML (no JS needed).

Long version: We were trying to manually add <li> elements to the select2 contact log select box for the contact logs currently selected. A couple problems with that approach: a) we couldn't get the JS to display correctly when our code ran, but it worked just fine in the Chrome console and b) everytime the select box is updated, any manual entries that have been added are removed. This means that our approach wasn't a viable option, so it makes more sense to just have the already selected contact logs hard-coded as HTML, because the user isn't able to remove them anyway (in this story).

Time spent: 0.4hr (20 min @ 50min/hr)

 -- Cam